Format source to enable unittesting and fix source bugs.
